* Switch random back to 32 bits. It's enough (tm).

+++ src/random.h
@@ -16,6 +16,6 @@
  *  random.h: pseudorandom number generator
  */
 
-void zzuf_srand(uint64_t);
-uint64_t zzuf_rand(uint64_t);
+void zzuf_srand(uint32_t);
+uint32_t zzuf_rand(uint32_t);
